Feature/features spec

+++ app/controllers/answers_controller.rb
@@ -1,11 +1,17 @@
+# frozen_string_literal: true
 class AnswersController < ApplicationController
-  expose :answer, build: ->(answer_params, _scope) { question.answers.new(answer_params) }
+  before_action :authenticate_user!, only: %i[new create destroy]
+
+  expose :question, -> { Question.find(params[:question_id]) }
+  expose :answers, -> { question.answers.reject(&:new_record?) }
+  expose :answer,
+         build: ->(params, _scope) { question.answers.new(params.merge(user_id: current_user.id)) }
 
   def create
     if answer.save
-      redirect_to answer
+      redirect_to answer.question, notice: 'Answer create successfully'
     else
-      render :new
+      render 'questions/show'
     end
   end
 
@@ -18,7 +24,12 @@ class AnswersController < ApplicationController
   end
 
   def destroy
-    answer.destroy
+    if current_user.author_of?(answer)
+      answer.destroy
+      flash[:notice] = 'Answer delete successfully'
+    else
+      flash[:alert] = "You can not delete someone else's answer"
+    end
     redirect_to answer.question
   end
 
@@ -26,9 +37,5 @@ class AnswersController < ApplicationController
 
   def answer_params
     params.require(:answer).permit(:body)
-  end
-
-  def question
-    Question.find(params[:question_id])
   end
 end
