+++ app/controllers/questions_controller.rb
@@ -1,10 +1,15 @@
 class QuestionsController < ApplicationController
+  before_action :authenticate_user!, only: %i[new create destroy]
+
   expose :questions, -> { Question.all }
-  expose :question
+  expose :question,
+         build: ->(params, _scope) { current_user.questions.new(params) }
+  expose :answers, -> { question.answers }
+  expose :answer, -> { question.answers.new }
 
   def create
     if question.save
-      redirect_to question
+      redirect_to question, notice: 'Question create successfully'
     else
       render :new
     end
@@ -19,7 +24,12 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
+    if current_user.author_of?(question)
+      question.destroy
+      flash[:notice] = 'Question delete successfully'
+    else
+      flash[:alert] = "You can not delete someone else's question"
+    end
     redirect_to questions_path
   end
